Rewrite UIImage extensions to use CoreGraphics APIs only
